Use ComputeContext backend for one-shot APIs
